Added table format rules
Added Hash table to define format rules.

Key is the column name, and value is an array of rules to check. the key is the expression (with the actual cell value prepended when generating HTML), second value contains the scops of highlighting (Cell, Col, or Row), comma, then the HTML tag attribute and value.