Updated 73 to allow for VMs stored on multiple datastores
Previously if a VM was stored on multiple datastores the plugin would
error out due to not being able to compare MemoryMB to a string
(FreeSpaceMB was returned as "AAAA BBBBB", A being the size for one
datastore B the size of the other). The plugin now finds the datastore
the VMX file is stored on and compares the memory to that.