22 Checking VM Hardware Version - FIlter VMs
I've added a filter, similar to the one used in the datastore
information plugin, that will exclude vms from this report. In my case
it is mainly used for DSVAs and vShield appliances which will remain on
HWversion 7.

Have tested on my vCenter reports and it seems to be working fine.