Update 02 Snapshot Information.ps1
Added filter options to TaskFilterSpec to filter not only startime but also entity. A check for entity name is done in line 91 which is now redundant as we are filtering on the entity.

In our environment this in some cases allowed us to do one iteration in the while loop instead of over 700. Speeding up the script from taking >5 hours to only take minutes.