Update 105 Host Profile Compliance.ps1
Minor change to lines 25 and 27. If only one host comes up as conforming to a host profile in a cluster the $VMHosts variable  does not return as an array when the failures are added in line 29. These are then dropped from the report. By forcing the assignment to an array in lines 25 and 27 the add will always work.