Modified to support command line parameters Server and DataCenter
Modified script to allow features of Server and DataCenter parameters being passed as command line options in vCheck.  Removed defining $Server from this script to allow vCheck to define Server.  Added test to check if DataCenter was defined and if DataCenter is defined add the -Location "$DataCenter" to the command lines of Get-VM, Get-VMHost, Get-Cluster and Get-Datastore.