Add command line parameters Server and DataCenter
Added command line parameters to vCheck to allow a Server and/or DataCenter to be passed in.

The Server parameter would allow environments with multiple servers to run the report against a single server and not need to create multiple job files or other work around methods.  This would require removing $Server from the Connection Plugin as well.

The DataCenter parameter would allow the report to filter items for a specific DataCenter.  This would require changes to the Connection Plugin to check for DataCenter being set and then adding the -Location "$DataCenter" to commands that accept it.

Defaults
    .\vCheck.ps1

Run for non-default server vcServer02
     .\vCheck.ps1 -Server vcServer02

Filter for DataCenter DC01
     .\vCheck.ps1 -DataCenter DC01

Filter for DataCenter DC01 on vcServer02
     .\vCheck.ps1 -Server vcServer02 -DataCenter DC01