Added functionality for multiple vCenters
Moved the $Server variable from the Connection plugin to the GlobalVariables file, and changed variable name to $Servers.

In the main vCheck script, I wrapped the entire processing portion of the script in a ForEach loop for the $Servers variable, and create the $Server variable (so no other code needs to be changed).  I re-import the GlobalVariables.ps1 script in each iteration to re-assign variables (such as email title, etc.) for dynamic assignment based on the $Server variable.

Now the $Servers variable, which is configured at the same point when running setup will be a comma separated list of vCenter servers.  Tested, and received 1 report per vCenter.