Update 48 Find VM Disk Format.ps1 - Filter Disktype
Adding '-DiskType Flat' to Get-Harddisk will improve performance as the data will be filtered in vCenter before being returned to the cmdlet.  As a minor bonus, it will also reduce output noise to the screen.  An error is generated for any non-flat disks because they don't have the StorageFormat property.